refactor: change flex factor to u32 (#32)

+++ src/layout/horizontal.rs
@@ -230,7 +230,7 @@ impl HorizontalLayout {
     }
 
     /// Sum up all the flex factors
-    fn flex_total(&self) -> u8 {
+    fn flex_total(&self) -> u32 {
         // TODO: if max width is set should we exclude
         // from flex factor?
